removed scraping of legacy mixnodes in NS api
