--- nym-node-status-api/nym-node-status-api/.sqlx/query-a19da2073bc691fe5cd2119a9527c16d6a8806ba50dcb759ac705b1a19288ca9.json
@@ -1,44 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n        SELECT\n            date_utc as \"date_utc!\",\n            SUM(total_stake)::BIGINT as \"total_stake!: i64\",\n            SUM(packets_received)::BIGINT as \"total_packets_received!: i64\",\n            SUM(packets_sent)::BIGINT as \"total_packets_sent!: i64\",\n            SUM(packets_dropped)::BIGINT as \"total_packets_dropped!: i64\"\n        FROM (\n            SELECT\n                date_utc,\n                n.total_stake,\n                n.packets_received,\n                n.packets_sent,\n                n.packets_dropped\n            FROM nym_node_daily_mixing_stats n\n            UNION ALL\n            SELECT\n                m.date_utc,\n                m.total_stake,\n                m.packets_received,\n                m.packets_sent,\n                m.packets_dropped\n            FROM mixnode_daily_stats m\n            LEFT JOIN nym_node_daily_mixing_stats ON m.mix_id = nym_node_daily_mixing_stats.node_id\n            WHERE nym_node_daily_mixing_stats.node_id IS NULL\n        )\n        GROUP BY date_utc\n        ORDER BY date_utc ASC\n        ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "date_utc!",
-        "type_info": "Varchar"
-      },
-      {
-        "ordinal": 1,
-        "name": "total_stake!: i64",
-        "type_info": "Int8"
-      },
-      {
-        "ordinal": 2,
-        "name": "total_packets_received!: i64",
-        "type_info": "Int8"
-      },
-      {
-        "ordinal": 3,
-        "name": "total_packets_sent!: i64",
-        "type_info": "Int8"
-      },
-      {
-        "ordinal": 4,
-        "name": "total_packets_dropped!: i64",
-        "type_info": "Int8"
-      }
-    ],
-    "parameters": {
-      "Left": []
-    },
-    "nullable": [
-      null,
-      null,
-      null,
-      null,
-      null
-    ]
-  },
-  "hash": "a19da2073bc691fe5cd2119a9527c16d6a8806ba50dcb759ac705b1a19288ca9"
-}
