+++ nym-node-status-api/nym-node-status-api/src/db/queries/mod.rs
@@ -1,7 +1,6 @@
 mod gateways;
 mod gateways_stats;
 mod misc;
-mod mixnodes;
 mod nym_nodes;
 mod packet_stats;
 pub(crate) mod scraper;
@@ -14,10 +13,10 @@ pub(crate) use gateways::{
 };
 pub(crate) use gateways_stats::{delete_old_records, get_sessions_stats, insert_session_records};
 pub(crate) use misc::insert_summaries;
-pub(crate) use mixnodes::{get_all_mixnodes, get_bonded_mix_ids, get_daily_stats, update_mixnodes};
 pub(crate) use nym_nodes::{
-    get_all_nym_nodes, get_bonded_node_description, get_described_bonded_nym_nodes,
-    get_described_node_bond_info, get_node_self_description, update_nym_nodes,
+    get_all_nym_nodes, get_bonded_node_description, get_daily_stats,
+    get_described_bonded_nym_nodes, get_described_node_bond_info, get_node_self_description,
+    update_nym_nodes,
 };
 pub(crate) use packet_stats::{
     batch_store_packet_stats, get_raw_node_stats, insert_daily_node_stats_uncommitted,
